% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dataprep.R
\name{Dataprep}
\alias{Dataprep}
\title{Normalize count matrix}
\usage{
Dataprep(sc_data, ref_data, peak_selection = 0.03)
}
\arguments{
\item{sc_data}{Input single cell ATAC sequence count matrix, rows refer to features/regions, columns refer to cells.}

\item{ref_data}{Input reference data, rows refer to observations, columns refer to features/regions.}

\item{peak_selection}{Parameter for peak selection, default value is peak_selection = 0.03; if keep all features, use peak_selection = 0.}
}
\value{
A list containing following components:
\item{sc_data}{Normalized scATAC-seq count matrix with selected features.}
\item{ref_data}{Normalized reference count matrix with selected features.}
}
\description{
This funciton is optional for normalization of runRA3's input count matrix. It consists of two steps:
}
\details{
1. Use TF-IDF for matrix normailization,

2. Select peaks for rough dimension reduction.
}
