% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RA3_motif.R
\name{getClusterSpecificPvalue}
\alias{getClusterSpecificPvalue}
\title{getClusterSpecificPvalue}
\usage{
getClusterSpecificPvalue(
  data,
  cluster,
  offset,
  landmark = NULL,
  maxiter = 1000,
  thresMLE = 10^-3,
  thresMAP = 10^-5,
  quiet = FALSE
)
}
\value{
A list of beta, p-value and sigmas.
}
\description{
getClusterSpecificPvalue
}
