% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RA3_motif.R
\name{getContrast}
\alias{getContrast}
\title{getContrast}
\usage{
getContrast(nCluster)
}
\value{
A list containing following items:
\item{constrasts}{contrasts needed to calculate the p-value}
\item{constrastsCluster}{clusters for contrasts}
}
\description{
getContrast
}
