% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RA3_clustering_test.R
\name{RA3_clustering}
\alias{RA3_clustering}
\title{Cell clustering based on the latent features output by RA3}
\usage{
RA3_clustering(H, n_cluster)
}
\arguments{
\item{H}{the extracted features H output by runRA3, colomns referring to cells}

\item{n_cluster}{the number of clusters}
}
\value{
\item{Seurat_louvain}{vector containing cell cluster assignment}
}
\description{
Implement louvain clustering based on the output of RA3, using package Seurat.
}
\examples{
result <- RA3_clustering(H, n_cluster)
}
