% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RA3_EM.R
\name{RA3_EM}
\alias{RA3_EM}
\title{EM algorithm for RA3}
\usage{
RA3_EM(Y, normalize, ttest, K1, K2, K3, Gamma, A, W, sigma_s)
}
\arguments{
\item{Y}{Input single cell count matrix.}

\item{normalize}{A logical value indicating whether the output W should be normalized through columns. Corresponding scale would be multiplied onto H if \emph{normalize} is TRUE.}

\item{ttest}{A logical value indicating whether a one sample ttest should be done for output H2 components.}

\item{K1, K2, K3}{Number of components consisted in RA3 model.}

\item{Gamma}{Initial matrix of parameter matrix \eqn{\Gamma}.}

\item{A}{Initial matrix of precision matrix A.}

\item{W}{Initial matrix of parameter matrix W, a warm start is recommended in \code{\link{runRA3}}.}

\item{sigma_s}{Intial value of parameter \eqn{\sigma^2}, a warm start is recommended in \code{\link{runRA3}}.}
}
\value{
A list containing the following components:
\item{H}{the estimated latent variable H.}
\item{W}{the estimated matrix of parameter matrix W.}
\item{Beta}{the estimated covariance parameter vector \eqn{\beta}.}
\item{Gamma}{the estimated indicator matrix \eqn{\Gamma}.}
\item{A}{the estimated precision matrix A.}
\item{sigma_s}{the estimated \eqn{\sigma^2}.}
\item{lgp}{the largest log posterior value when EM algorithm converges.}
}
\description{
This function is implementing an EM algorithm to estimate parameters of RA3 model.
}
