% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RA3_motif.R
\name{RA3_motif}
\alias{RA3_motif}
\title{Motif enrichment analysis based on the latent features output by RA3}
\usage{
RA3_motif(
  peaks,
  label_true,
  label_est,
  sc_mat,
  cluster_peak_num = 1000,
  motif_num = 50
)
}
\arguments{
\item{peaks}{a vector containing peak information of the scCAS data, with elements formatted as 'chr1_24516561_24517061'}

\item{label_true}{true cell labels}

\item{label_est}{the estimated cluster labels, starting from 1}

\item{sc_mat}{scCAS count matrix}

\item{cluster_peak_num}{number of peaks selected for each cluster, default value is 1000}

\item{motif_num}{number of motifs shown in the plot, default value is 50}
}
\value{
A list containing following items:
\item{metaDataInd}{cell index grouped by clustering results and cell labels}
\item{top_devs}{derivation scores estimated using chromVAR}
}
\description{
The output of RA3 can also be implemented in motif enrichment analysis. This function uses the extracted features from RA3 for motif analysis though package chromVAR .
}
\examples{
result <- RA3_motif(peaks, label_true, label_est, sc_mat,cluster_peak_num=1000, motif_num=50)

}
