% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RA3_pcatsne.R
\name{RA3_pcatsne}
\alias{RA3_pcatsne}
\title{Dimension reduction using PCA and t-SNE}
\usage{
RA3_pcatsne(sc_data_nml, rand_seed = 23)
}
\arguments{
\item{sc_data_nml}{TF-IDF normalized scCAS data, the output of function Dataprep}
}
\value{
\item{pca_tsne}{tsne representation for component scores of the TF-IDF normalized scCAS data}
}
\description{
Implement PCA on the TF-IDF normalized scCAS data, then deploy t-SNE on the component scores.
}
\examples{
result <- RA3_pcatsne(sc_data_nml)
}
