% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RA3_RefProj.R
\name{RA3_RefProj}
\alias{RA3_RefProj}
\title{Dimension reduction using reference projection approach}
\usage{
RA3_RefProj(ref_data_nml, sc_data_nml, rand_seed = 23)
}
\arguments{
\item{sc_data_nml}{TF-IDF normalized scCAS data, the output of function Dataprep}
}
\value{
\item{bulk_tsne}{tsne representation for projected scCAS data}
}
\description{
This function deploys a reference projection approach on the TF-IDF normalized scCAS data, then uses t-SNE for further dimension reduction.
}
\examples{
result <- RA3_RefProj(ref_data_nml,sc_data_nml,rand_seed = 2020)
}
