% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RA3_TrajInf.R
\name{RA3_TrajInf}
\alias{RA3_TrajInf}
\title{Trajectory inference based on the latent features output by RA3}
\usage{
RA3_TrajInf(DRres_mat, cluster_assign, rand_seed = 23)
}
\arguments{
\item{DRres_mat}{the extracted features H output by runRA3, colomns referring to cells}

\item{cluster_assign}{the estimated cluster labels}

\item{rand_seed}{random seed set for tsne plot, default value is set as 23}
}
\value{
A list  containing following items:
\item{tsne}{matrix containing tsne representation, calculated using random seed 'rand_seed'}
\item{sds.new}{A SlingshotDataSet containing curves. This contains the curves representing the estimated cell lineages.}
}
\description{
The output of RA3 can also be implemented in trajectory inference. This function uses the extracted features from RA3 for trajectory inference by slingshot.
}
\examples{
result <- RA3_TrajInf(DRres_mat,cluster_assign,rand_seed = 23)
}
