% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRA3.R
\name{runRA3}
\alias{runRA3}
\title{Run RA3 for integrative analysis of scATAC-seq data}
\usage{
runRA3(sc_data, ref_data, K2 = 5, K3 = 5)
}
\arguments{
\item{sc_data}{scATAC-seq count matrix, the rows should refer to features/regions and columns refer to cells.}

\item{ref_data}{reference data matrix, the columns should refer to features/regioins and rows refer to observations.}

\item{K2}{the number of components in RA3's second part, the default value is K2 = 5.}

\item{K3}{the number of components in RA3's third part, the default value is K3 = 5.}
}
\value{
A list containing the following components:
\item{H}{the extracted latent features H.}
\item{W}{the estimated matrix of parameter matrix W.}
\item{Beta}{the estimated covariance parameter vector \eqn{\beta}.}
\item{Gamma}{the estimated indicator matrix \eqn{\Gamma}.}
\item{A}{the estimated precision matrix A.}
\item{sigma_s}{the estimated \eqn{\sigma^2}.}
\item{lgp}{the largest log posterior value when EM algorithm converges.}
}
\description{
RA3 refers to Reference-guided Approach for the Analysis of scATAC-seq data. It can simultaneously incorporate shared biological variation from reference data and identify distinct subpopulations, and thus achieves superior performance to existing methods in comprehensive experiments.
}
\examples{
result <- runRA3(sc_example, reference_example)
result <- runRA3(sc_example, reference_example, 10, 5)
}
